refactor scheduler pr
