+++ scripts/run_train.py
@@ -19,35 +19,7 @@ from torch_ema import ExponentialMovingAverage
 import mace
 from mace import data, modules, tools
 from mace.tools import torch_geometric
-from mace.tools.scripts_utils import create_error_table, get_dataset_from_xyz
-
-class LrScheduler:
-    def __init__(self, optimizer, args) -> None:
-        self.scheduler = args.scheduler
-        if args.scheduler == "ExponentialLR":
-            self.lr_scheduler = torch.optim.lr_scheduler.ExponentialLR(
-                optimizer=optimizer, gamma=args.lr_scheduler_gamma
-                )
-        elif args.scheduler == "ReduceLROnPlateau":
-            self.lr_scheduler = torch.optim.lr_scheduler.ReduceLROnPlateau(
-                optimizer=optimizer,
-                factor=args.lr_factor,
-                patience=args.scheduler_patience,
-                )
-        else:
-            raise RuntimeError(f"Unknown scheduler: '{args.scheduler}'")
-
-    def step(self, metrics=None, epoch=None):
-        if self.scheduler == "ExponentialLR":
-            self.lr_scheduler.step(epoch=epoch)
-        elif self.scheduler == "ReduceLROnPlateau":
-            self.lr_scheduler.step(metrics=metrics, epoch=epoch)
-
-    def state_dict(self):
-        return {key: value for key, value in self.lr_scheduler.__dict__.items() if key != 'optimizer'}
-
-    def load_state_dict(self, state_dict):
-        self.lr_scheduler.__dict__.update(state_dict)
+from mace.tools.scripts_utils import create_error_table, get_dataset_from_xyz, LRScheduler
 
 
 def main() -> None:
@@ -397,7 +369,7 @@ def main() -> None:
 
     logger = tools.MetricsLogger(directory=args.results_dir, tag=tag + "_train")
 
-    lr_scheduler = LrScheduler(optimizer, args)
+    lr_scheduler = LRScheduler(optimizer, args)
 
     swa: Optional[tools.SWAContainer] = None
     swas = [False]
